updated README
