+++ tests/unit/examples/test_notebooks_gpu.py
@@ -124,3 +124,10 @@ def test_dkn_quickstart(notebooks, output_notebook, kernel_name):
         kernel_name=kernel_name,
         parameters=dict(epochs=1, batch_size=500),
     )
+
+
+@pytest.mark.notebooks
+@pytest.mark.gpu
+def test_sasrec_single_node_runs(notebooks, output_notebook, kernel_name):
+    notebook_path = notebooks["sasrec_quickstart"]
+    pm.execute_notebook(notebook_path, output_notebook, kernel_name=kernel_name)
